Refactor app structure with templates and blueprints
