+++ src/routes/__init__.py
@@ -1,1 +1,7 @@
-# Ce fichier permet de marquer le dossier comme un package Python
+from .song_routes import song_bp
+from .main import main_bp
+from .recordings import recordings_bp
+from .goals import goals_bp
+from .songsterr import songsterr_bp
+
+__all__ = ['song_bp', 'main_bp', 'recordings_bp', 'goals_bp', 'songsterr_bp']
