+++ src/static/css/style.css
@@ -1,0 +1,9 @@
+/* Custom styles */
+body {
+    font-family: Arial, sans-serif;
+    margin-top: 30px;
+    text-align: center;
+}
+.navbar {
+    margin-bottom: 20px;
+}
